+++ examples/experimental/dagster-airlift/dagster_airlift/core/__init__.py
@@ -1,10 +1,10 @@
 from ..proxied_state import load_proxied_state_from_yaml as load_proxied_state_from_yaml
 from .basic_auth import BasicAuthBackend as BasicAuthBackend
-from .dag_defs import (
-    dag_defs as dag_defs,
-    task_defs as task_defs,
-)
 from .load_defs import (
     AirflowInstance as AirflowInstance,
     build_defs_from_airflow_instance as build_defs_from_airflow_instance,
+)
+from .top_level_dag_def_api import (
+    dag_defs as dag_defs,
+    task_defs as task_defs,
 )
